hotfix: deleteCard fixed

+++ controllers/cards.js
@@ -30,10 +30,17 @@ module.exports.createCard = async (req, res) => {
 module.exports.deleteCard = async (req, res) => {
   try {
     const { cardId } = req.params;
+    const card = await Card.findById(cardId).populate('owner');
+
+    if(!card) {
+      return res.status(ERROR_CODE_NOT_FOUND).send({ message: "Карточка с таким id не найдена" });
+    }
+
     await Card.findByIdAndDelete(cardId);
-    return res.send({ message: "Карточка удалена" });
+
+    return res.send(card);
   } catch (error) {
-    if ((error.message === "NotFound")) {
+    if (error.message === "NotFound") {
       return res.status(ERROR_CODE_NOT_FOUND).send({ message: "Карточка с таким id не найдена" });
     }
     return res.status(ERROR_CODE_SERVER_ERROR).send({ message: "Ошибка на стороне сервера", error: error.message });
@@ -46,7 +53,11 @@ module.exports.likeCard = async (req, res) => {
       req.params.cardId,
       { $addToSet: { likes: req.user._id } },
       { new: true, runValidators: true },
-    )
+    );
+
+    if (!likedCard) {
+      return res.status(ERROR_CODE_NOT_FOUND).send({ message: "Карточка с таким id не найдена" });
+    }
 
     return res.send(likedCard);
   } catch (error) {
@@ -65,10 +76,15 @@ module.exports.dislikeCard = async (req, res) => {
       { $pull: { likes: req.user._id } },
       { new: true, runValidators: true },
     )
+
+    if (!dislikedCard) {
+      return res.status(ERROR_CODE_NOT_FOUND).send({ message: "Карточка с таким id не найдена" });
+    }
+
     return res.send(dislikedCard);
   } catch (error) {
-    if ((error.message === "NotFound")) {
-      return res.status(ERROR_CODE_NOT_FOUND).send({ message: "Карточка с таким id не найдена" });
+    if ((error.name === "CastError" || error.name === "ValidationError")) {
+      return res.status(ERROR_CODE_VALIDATION).send({ message: "Карточка с таким id не найдена", ...error });
     }
     return res.status(ERROR_CODE_SERVER_ERROR).send({ message: 'Произошла ошибка' });
   }
